+++ scripts/Orders.js
@@ -1,4 +1,4 @@
-import { getMetals, getOrders, getSizes, getStyles } from "./database.js"
+import { getMetals, getOrders, getSizes, getStyles } from "./dataAccess.js"
 
 const metals = getMetals()
 const sizes = getSizes()
