+++ scripts/Types.js
@@ -1,4 +1,4 @@
-import { setType } from "./database.js"
+import { setType } from "./dataAccess.js"
 
 document.addEventListener("change",(event) => { 
     if (event.target.name === "type") {
